update - fix app structure, rename components
