--- essentialism/src/App.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
-import UserForm from "./UserForm"
-import ValueList from "./ValueList"
-import {BrowserRouter as Router, Route} from "react-router-dom";
-import UserFormOne from './UserFormOne';
-
-function App() {
-  return (
-    <div className="App">
-      <UserForm/>
-      {/* <ValueList/> */}
-    </div>
-  );
-}
-
-export default App;
